raspberry/message_subscriber/requirements.txt
pika==1.0.1
RPi.GPIO==0.6.5
